OmniSport_AI/requirements.txt
streamlit
pandas 
requests 
python-dotenv
openai 
matplotlib
seaborn 
plotly